exercises/01-webpage-summarizer/requirements.txt
python-dotenv
requests
beautifulsoup4
openai
pytest